04-유상민/비트컴퓨터_프로젝트_발표_최종/requirements.txt
streamlit
pandas
chromadb==0.4.15
sentence-transformers
torch
python-dotenv
openai>=0.27.0
langchain-community








